chore: relive fast setup script

+++ scripts/fast-setup.sh
@@ -6,14 +6,18 @@
 # It should work, and should only be run ONCE.
 # If it doesn't work, a PR is welcome.
 
-cp .env.example .env
+[ ! -f .env ] && cp .env.example .env
 
+docker compose down -v
 docker compose up dev -d
 
 docker compose exec dev pnpm i
 docker compose exec dev pnpm build:all
-docker compose exec dev tsx scripts/download-data.ts 110-1 110-2 111-1
-docker compose exec dev tsx scripts/upload-data.ts data/110-1.json data/110-2.json data/111-1.json
+docker compose exec dev bash -c "cd packages/cli && npm link"
+docker compose exec dev bash -c "mkdir data && curl -L -o data/courses.json https://github.com/JacobLinCool/NTNU-Course-Crawler/releases/download/20221106/108-1.108-2.109-1.109-2.110-1.110-2.111-1.json"
+docker compose exec dev pnpm db:push
+docker compose exec dev bash -c "pnpm dev & sleep 5 && (echo 'unicourse' | unicourse login unicourse) && unicourse import data/courses.json && exit"
 
-echo You can now run \"pnpm dev\" and go to "http://localhost:8080/course/query?year=111&term=1&q=紀博文"
+echo Adminer is running at "http://localhost:8180"
+echo use \"pnpm dev\" to start the server at "http://localhost:8080"
 docker compose exec dev bash
